added gradient, always on
