+++ EZWorkout/Custom Views/ExerciseView.swift
@@ -11,6 +11,7 @@ class ExerciseView: UIView {
   @IBOutlet weak var labelExerciseWeight: UILabel!
   @IBOutlet weak var labelExerciseReps: UILabel!
   @IBOutlet weak var exerciseImage: UIImageView!
+  @IBOutlet weak var infoContainer: UIView!
   
   override init(frame: CGRect) {
     super.init(frame: frame)
@@ -32,6 +33,7 @@ class ExerciseView: UIView {
   func setUp(exerciseModel:ExerciseModel){
     
     //TODO adjust text presentation
+    infoContainer.layer.cornerRadius = 12
     exerciseImage.layer.cornerRadius = exerciseImage.frame.size.width / 2
     labelExerciseName.text = exerciseModel.name
     if let reps = exerciseModel.reps {
